Got everything compiling with visual studio IDE files, generated with cmake Added visual studio release files Added visual studio roadrunner installer Tested quickly: compilation and running seems to work
Some progress with python wrapper, but not clear how to incorporate it into CMake
Tested Delphi project, compiled,ran on my machine after commenting

//  madExcept,
//  madLinkDisAsm,
//  madListHardware,
//  madListProcesses,
//  madListModules,

+++ trunk/src/rrModelFromC.cpp
@@ -318,7 +318,7 @@ bool ModelFromC::SetupDLLData()
         Log(lDebug)<<"Failed to assign to mC";
     }
 
-    cSize       = (double*) GetProcAddress((HMODULE) mDLLHandle, "_cSize");
+    cSize       = (int*) GetProcAddress((HMODULE) mDLLHandle, "_cSize");
     if(!cSize)
     {
         Log(lDebug)<<"Failed to assign to cSize";
